refactor(server): extract command wake orchestration
